refactored variants
