+++ gateway/src/variant/chat_completion.rs
@@ -14,21 +14,15 @@ use crate::inference::types::{
     InputMessageContent, ModelInferenceRequest, ModelInferenceRequestJsonMode,
     ModelInferenceResponseWithMetadata, RequestMessage, Role,
 };
-use crate::jsonschema_util::DynamicJSONSchema;
 use crate::minijinja_util::TemplateConfig;
-use crate::tool::{create_dynamic_implicit_tool_config, ToolCallConfig};
+use crate::tool::create_dynamic_implicit_tool_config;
+use crate::variant::JsonMode;
 use crate::{
     inference::types::{InferenceResult, InputMessage},
     model::ModelConfig,
 };
 
-#[derive(Debug, Deserialize)]
-#[serde(tag = "type")]
-#[serde(rename_all = "snake_case")]
-#[serde(deny_unknown_fields)]
-pub enum VariantConfig {
-    ChatCompletion(ChatCompletionConfig),
-}
+use super::{InferenceConfig, ModelUsedInfo, Variant};
 
 #[derive(Debug, Default, Deserialize)]
 #[serde(deny_unknown_fields)]
@@ -43,146 +37,6 @@ pub struct ChatCompletionConfig {
     pub seed: Option<u32>,
     #[serde(default)]
     pub json_mode: JsonMode, // Only for JSON functions, not for chat functions
-}
-
-/// This type is used to determine how to enforce JSON mode for a given variant.
-/// Variants represent JSON mode in a slightly more abstract sense than ModelInferenceRequests, as
-/// we support coercing tool calls into JSON mode.
-/// This is represented as a tool config in the
-#[derive(Debug, Default, Deserialize, PartialEq)]
-#[serde(rename_all = "snake_case")]
-pub enum JsonMode {
-    Off,
-    #[default]
-    On,
-    Strict,
-    ImplicitTool,
-}
-
-/// Maps to the subset of Config that applies to the current inference request.
-/// It doesn't take into account inference-time overrides (e.g. dynamic tools).
-pub struct InferenceConfig<'a> {
-    pub tool_config: Option<ToolCallConfig>,
-    pub templates: &'a TemplateConfig<'a>,
-    pub dynamic_output_schema: Option<DynamicJSONSchema>,
-}
-
-pub struct ModelUsedInfo<'a> {
-    pub model_name: &'a str,
-    pub model_provider_name: &'a str,
-}
-
-pub trait Variant {
-    async fn infer<'a, 'request>(
-        &'a self,
-        input: &Input,
-        models: &'a HashMap<String, ModelConfig>,
-        function: &'a FunctionConfig,
-        inference_config: &'request InferenceConfig<'request>,
-        client: &'request Client,
-        inference_params: &mut InferenceParams,
-    ) -> Result<InferenceResult<'a>, Error>;
-
-    async fn infer_stream<'request>(
-        &'static self,
-        input: &Input,
-        models: &'static HashMap<String, ModelConfig>,
-        function: &'static FunctionConfig,
-        inference_config: &'request InferenceConfig<'request>,
-        client: &'request Client,
-        inference_params: &mut InferenceParams,
-    ) -> Result<
-        (
-            InferenceResultChunk,
-            InferenceResultStream,
-            ModelUsedInfo<'static>,
-        ),
-        Error,
-    >;
-}
-
-impl VariantConfig {
-    pub fn weight(&self) -> f64 {
-        match self {
-            VariantConfig::ChatCompletion(params) => params.weight,
-        }
-    }
-    pub fn system_template(&self) -> Option<&PathBuf> {
-        match self {
-            VariantConfig::ChatCompletion(params) => params.system_template.as_ref(),
-        }
-    }
-
-    pub fn user_template(&self) -> Option<&PathBuf> {
-        match self {
-            VariantConfig::ChatCompletion(params) => params.user_template.as_ref(),
-        }
-    }
-
-    pub fn assistant_template(&self) -> Option<&PathBuf> {
-        match self {
-            VariantConfig::ChatCompletion(params) => params.assistant_template.as_ref(),
-        }
-    }
-}
-
-impl Variant for VariantConfig {
-    async fn infer<'a, 'request>(
-        &'a self,
-        input: &Input,
-        models: &'a HashMap<String, ModelConfig>,
-        function: &'a FunctionConfig,
-        inference_config: &'request InferenceConfig<'request>,
-        client: &'request Client,
-        inference_params: &mut InferenceParams,
-    ) -> Result<InferenceResult<'a>, Error> {
-        match self {
-            VariantConfig::ChatCompletion(params) => {
-                params
-                    .infer(
-                        input,
-                        models,
-                        function,
-                        inference_config,
-                        client,
-                        inference_params,
-                    )
-                    .await
-            }
-        }
-    }
-
-    async fn infer_stream<'request>(
-        &'static self,
-        input: &Input,
-        models: &'static HashMap<String, ModelConfig>,
-        function: &'static FunctionConfig,
-        inference_config: &'request InferenceConfig<'request>,
-        client: &'request Client,
-        inference_params: &mut InferenceParams,
-    ) -> Result<
-        (
-            InferenceResultChunk,
-            InferenceResultStream,
-            ModelUsedInfo<'static>,
-        ),
-        Error,
-    > {
-        match self {
-            VariantConfig::ChatCompletion(params) => {
-                params
-                    .infer_stream(
-                        input,
-                        models,
-                        function,
-                        inference_config,
-                        client,
-                        inference_params,
-                    )
-                    .await
-            }
-        }
-    }
 }
 
 impl ChatCompletionConfig {
@@ -394,10 +248,10 @@ mod tests {
     use crate::inference::providers::common::get_temperature_tool_config;
     use crate::inference::providers::dummy::{DummyProvider, DUMMY_JSON_RESPONSE_RAW};
     use crate::inference::types::{ContentBlockOutput, Usage};
-    use crate::jsonschema_util::JSONSchemaFromPath;
+    use crate::jsonschema_util::{DynamicJSONSchema, JSONSchemaFromPath};
     use crate::minijinja_util::tests::get_test_template_config;
     use crate::model::ProviderConfig;
-    use crate::tool::ToolChoice;
+    use crate::tool::{ToolCallConfig, ToolChoice};
     use crate::{
         error::Error,
         inference::{
